requirements.txt
asgiref==3.5.1
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bandit==1.7.4
black==22.3.0
cached-property==1.5.2
click==8.1.3
colorama==0.4.4
coverage==6.4
decorator==5.1.1
Django==3.2
django-extensions==3.1.5
django-url-filter==0.3.15
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.1
enum-compat==0.0.3
executing==0.8.3
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.27
inflection==0.5.1
iniconfig==1.1.1
ipython==8.3.0
jedi==0.18.1
jsonschema==4.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pbr==5.9.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
stack-data==0.2.0
stevedore==3.5.0
toml==0.10.2
tomli==2.0.1
traitlets==5.2.1.post0
uritemplate==4.1.1
wcwidth==0.2.5
